requirement.txt
Django>=3.0.4
django-cors-headers>=3.2.1
djangorestframework>=3.11.0
djangorestframework-jwt>=1.11.0
gunicorn>=20.0.4
idna>=2.9
Pillow>=7.0.0
psycopg2>=2.8.4
PyJWT>=1.7.1
python-http-client>=3.2.6
pytz>=2019.3
requests>=2.23.0
sendgrid>=6.2.0
simplejson>=3.17.0
google-play-scraper==0.0.2.6
beautifulsoup4==4.9.1
mysqlclient==1.4.6